Fix access token grabbing, more auth0 seeding
Fixes [AS-273](https://cs4500-jira.ccs.neu.edu/browse/AS-273) by only
grabbing the access token for Auth0 after the previous one has expired.

Also ensures that Auth0 is in expected state before running the
auth0_utils tests so that it does not error out and break the build.

This is more work for
[AS-60](https://cs4500-jira.ccs.neu.edu/browse/AS-60)

Lints.